+++ src/PokeGame.Core/Moves/Models/CreateOrReplaceMovePayload.cs
@@ -12,9 +12,9 @@ public record CreateOrReplaceMovePayload
   public string? Name { get; set; }
   public string? Description { get; set; }
 
-  public byte? Accuracy { get; set; }
-  public byte? Power { get; set; }
-  public byte PowerPoints { get; set; }
+  public int? Accuracy { get; set; }
+  public int? Power { get; set; }
+  public int PowerPoints { get; set; }
 
   public void Validate() => new Validator().ValidateAndThrow(this);
